# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUser:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
  models.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.UserModel:
    properties:
      _id:
        type: string
      bio:
        type: string
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      following:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
      password:
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
host: localhost:5005
info:
  contact: {}
  description: This is Swagger docs for rest api goland fiber
  title: Fiber Goland Mongo Grpc Websocet etc..
  version: "1.0"
paths:
...
  /user/signin:
    post:
      consumes:
      - application/json
      description: Login an ew user by providing email, password
      parameters:
      - description: user login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Login a new user
      tags:
      - Authentication
  /user/signup:
    post:
      consumes:
      - application/json
      description: Register an ew user by providing email, password, first name, last
        name
      parameters:
      - description: user register details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"